pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "similarity-relation"
version = "0.1.0"
description = "Multi-view paper and author similarity experiments for citation graphs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "pandas>=2.1",
  "scikit-learn>=1.4",
  "pyyaml>=6.0",
  "pyarrow>=15.0",
  "networkx>=3.2",
  "tqdm>=4.66",
  "typer>=0.12",
  "rich>=13.7"
]

[project.optional-dependencies]
text = [
  "sentence-transformers>=3.0"
]
dev = [
  "pytest>=8.0",
  "jupyter>=1.0"
]

[project.scripts]
similarity-relation = "mvsim.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
